Move components to Molecules folder
