+++ packages/react/src/Components/Atoms/index.js
@@ -2,7 +2,5 @@ export { default as Button } from './Button';
 export { default as CanvasImage } from './CanvasImage';
 export { default as CanvasRect } from './CanvasRect';
 export { default as Image } from './Image';
-export { default as MRZTableHeading } from './MRZTableHeading';
-export { default as MRZTableRow } from './MRZTableRow';
 export { default as Video } from './Video';
 export { default as Tag } from './Tag';
